delete from app.js routes, made route.js ,add contoroller, model,setup
